gs and ifta runtime benchmarking script, cpu and gpu separated
